feat(h5): support resize fix #10967
